Added sftp data reader
