+++ radarpipeline/io/__init__.py
@@ -1,4 +1,4 @@
 from radarpipeline.io.abc import DataReader, SchemaReader
 from radarpipeline.io.reader import AvroSchemaReader, SparkCSVDataReader, Reader
-from radarpipeline.io.reader import SftpDataReader
+from radarpipeline.io.downloader import SftpDataReader
 from radarpipeline.io.writer import *
