+++ requirements.txt
@@ -16,4 +16,5 @@ pyspark-test==0.2.0
 pytest-cov
 pytest-mock
 PyGithub
+pygithub
 pygithub3
